Split Pregel V8 methods into their own file
