+++ arangod/V8Server/CMakeLists.txt
@@ -8,6 +8,7 @@ add_library(arango_v8server STATIC
   v8-collection.cpp
   v8-dispatcher.cpp
   v8-general-graph.cpp
+  v8-pregel.cpp
   v8-prototype-states.cpp
   v8-query.cpp
   v8-replicated-logs.cpp
