+++ arangod/V8Server/v8-vocbase.cpp
@@ -85,6 +85,7 @@
 #include "V8Server/v8-general-graph.h"
 #include "V8Server/v8-replicated-logs.h"
 #include "V8Server/v8-prototype-state.h"
+#include "V8Server/v8-pregel.h"
 #include "V8Server/v8-replication.h"
 #include "V8Server/v8-statistics.h"
 #include "V8Server/v8-users.h"
@@ -2240,6 +2241,7 @@ void TRI_InitV8VocBridge(v8::Isolate* isolate, v8::Handle<v8::Context> context,
   TRI_InitV8IndexArangoDB(isolate, ArangoNS);
 
   TRI_InitV8Collections(context, &vocbase, v8g, isolate, ArangoNS);
+  TRI_InitV8Pregel(isolate, ArangoNS);
   TRI_InitV8Views(*v8g, isolate);
   TRI_InitV8ReplicatedLogs(v8g, isolate);
   TRI_InitV8PrototypeStates(v8g, isolate);
